+++ api/modals/models.py
@@ -28,7 +28,6 @@ class StockMonitorModal(BaseModel):
     token: Any
     request: Any
     plaintext: bool = False
-    otp: Any = None
 
 
 class CameraIndexModal(BaseModel):