+++ api/squire/timeout_otp.py
@@ -3,7 +3,7 @@ from typing import NoReturn, Union
 
 from pydantic import EmailStr
 
-from api.settings import robinhood, stock_monitor_helper, surveillance
+from api.modals.settings import robinhood, stock_monitor_helper, surveillance
 
 
 def reset_robinhood(after: Union[int, float]) -> NoReturn:
@@ -32,13 +32,11 @@ def reset_stock_monitor(email_address: EmailStr, after: Union[int, float]) -> No
     start = time.time()
     while True:
         if start + after <= time.time():
-            del stock_monitor_helper.otp[email_address]
+            del stock_monitor_helper.otp_sent[email_address]
             break
-        if not stock_monitor_helper.otp[email_address]:  # Already set to None
+        if not stock_monitor_helper.otp_sent[email_address]:  # Already set to None
             break
         time.sleep(1)
-    if email_address in stock_monitor_helper.validated:
-        stock_monitor_helper.validated.remove(email_address)
 
 
 def reset_surveillance(after: Union[int, float]) -> NoReturn: