+++ doc_generator/conf.py
@@ -64,6 +64,10 @@ html_static_path = ['_static']
 # Reference: https://www.sphinx-doc.org/en/master/usage/extensions/autodoc.html#confval-autoclass_content
 autoclass_content = 'both'
 
+# Include private methods/functions
+# Reference: https://www.sphinx-doc.org/en/master/usage/extensions/autodoc.html#confval-autodoc_default_options
+autodoc_default_options = {"members": True, "undoc-members": True, "private-members": True}
+
 # Add support to mark down files in sphinx documentation
 # Reference: https://www.sphinx-doc.org/en/1.5.3/markdown.html
 source_suffix = {